move dashboard page to jobber/settings
